AP_NavEKF3: Prevent on ground air data estimation
This fixes a problem where if an airspeed sensor was fitted, wind speed estimation would run when the vehicle was on ground, causing invalid wind estimates to form.
This change adds the requirement to be in a forward flight mode which is reasonable because there are currently no COTS airspeed sensors available for SUAS which work for 360 deg AOS.
The less permissive isFlying check is also used which for takeoff in fixed wing modes will prevent on ground wind estimation caused by early arming.