AP_TECS: Don't reference throttle damping to height error
The current implementation calculates the potential energy rate setpoint used by the throttle damping term as proportional to height error, not the climb rate required by the demanded height profile. This means the damping term is responding to height error in addition to height and airspeed rate of change. This reduces damping and limits the amount of TECS_THR_DAMP that can be used.