AP_TECS: Fix height variable naming and add missing fade of height feedback
TECS uses two height measurements.
The primary one is height above home which is not named descriptively and is also incorrectly commented.
The other height is a height above touchdown point which is used only during flare.
The underlying issue was during flare, TECS was not correctly coordinating the switch between height above home and height above touchdown point datums.
This patch ensures the changeover is down smoothly and is applied to both the demanded and measured height.